% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor.R
\name{anchor}
\alias{anchor}
\title{Select anchor items}
\usage{
anchor(
  x,
  size,
  mean_diff = 0.01,
  sd_diff = 0.01,
  sd_spread = 1,
  max_secs = 5,
  sd_diff_adj = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{x}{A matrix of item responses.}

\item{size}{The number of anchor items.}

\item{mean_diff}{The maximum absolute difference between the mean of the
anchor test \eqn{p}-values and the mean of the total test \eqn{p}-values.
Default is \code{0.01}.}

\item{sd_diff}{The maximum absolute difference between the SD of the anchor
test \eqn{p}-values and the SD of the total test \eqn{p}-values multiplied
by \code{sd_spread}. Default is \code{0.01}.}

\item{sd_spread}{A value greater than 0 indicating the spread of the anchor
test \eqn{p}-values relative to the spread of the total test
\eqn{p}-values. \code{1} (default) indicates a minitest, and smaller values
indicate a semi-miditest or a miditest (Sinharay & Holland, 2007).}

\item{max_secs, sd_diff_adj}{The maximum number of seconds to search for a
solution. Default is \code{5}. If a solution is not found, then \code{sd_diff} is
increased by \code{sd_diff_adj} and a message is given.}

\item{verbose}{Logical indicating whether full output should be returned.
Default is \code{TRUE}. If \code{FALSE}, then only the anchor item indices are
returned.}
}
\value{
A list is returned with the following output.
\item{ind}{A vector of anchor item indices.}
\item{moments}{A matrix of the means and SDs of the total test \eqn{p}-values
and the anchor test \eqn{p}-values.}
\item{cor}{The correlation between the total test scores and the anchor test
scores.}
}
\description{
Select anchor items to represent the total test.
}
\examples{
# Extract data
list2env(nonequiv_groups_dat, envir = globalenv())

# Minitest
anchor(y, size = 30)

# Semi-miditest with a 75\% spread
anchor(y, size = 30, sd_spread = 0.75)

}
\references{
Sinharay, S., & Holland, P. W. (2007). Is it necessary to make anchor tests
mini-versions of the tests being equated or can some restrictions be relaxed?
\emph{Journal of Educational Measurement}, \emph{44}(3), 249--275.
}
\author{
Kylie N. Gorney \email{kyliengorney@gmail.com}
}
