% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot}
\alias{boot}
\title{Compute bootstrap standard errors}
\usage{
boot(object, rep = 1000, crit)
}
\arguments{
\item{object}{An object of class \code{eqt} or \code{multipleqt}.}

\item{rep}{The number of replications to be conducted. Default is \code{1000}.}

\item{crit}{A vector of scores to be used as the criterion when computing the
bias and RMSE. Defaults to the results of the original equating.}
}
\value{
An object of class \code{boot} is returned with the following output.
\item{scales}{A matrix of raw scores, criterion scores, mean equated scores,
standard errors of equating, bias, and RMSEs.}
\item{scores}{A matrix of raw scores, criterion scores, and mean equated
scores for each examinee.}
\item{mean_see}{A weighted average of the standard errors. Weights are chosen
using the frequencies of the observed scores (Kolen & Brennan, 2014).}
}
\description{
Compute bootstrap standard errors for one or more equatings.
}
\details{
Bootstrapping offers an alternative to computing analytic standard errors. In
this process, samples are repeatedly drawn from the original data with
replacement. Equating is then conducted on these samples, and results are
compared across replications.
}
\examples{
# Setup ---------------------------------------------------------------------

# Extract data
list2env(random_groups_dat, envir = globalenv())

# Identity equating
id <- eqt(x, y)

# Linear equating
li <- eqt(x, y, linear = "linear")

# Multiple equatings
id_li <- multipleqt(id, li)

# Compute bootstrap standard errors -----------------------------------------

# Use an eqt object
boot(li)

# Use a multipleqt object
boot(id_li)

}
\references{
Kolen, M. J., & Brennan, R. L. (2014). \emph{Test Equating, Scaling, and Linking:
Methods and Practices} (3rd ed.). Springer.
\url{https://doi.org/10.1007/978-1-4939-0317-7}
}
\seealso{
\code{\link[=eqt]{eqt()}}, \code{\link[=multipleqt]{multipleqt()}}
}
\author{
Kylie N. Gorney \email{kyliengorney@gmail.com}
}
