% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqt.R
\name{eqt}
\alias{eqt}
\title{Small-sample equating}
\usage{
eqt(
  x,
  y,
  xv = NULL,
  yv = NULL,
  linear = c("identity", "mean", "linear"),
  method = c("nominal weights", "tucker", "levine observed", "levine true", "chained"),
  circle = NULL,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{x, y}{Matrices of item responses from the new and old forms,
respectively.}

\item{xv, yv}{Numeric vectors containing indices of the anchor items, where
\code{xv} corresponds to \code{x}, and \code{yv} corresponds to \code{y}. If either is \code{NULL},
then the random groups design is assumed (see Details).}

\item{linear}{Character indicating the transformation to be applied.
\code{"identity"} (default) does not use any estimation, \code{"mean"} estimates the
intercept, and \code{"linear"} estimates the intercept and slope.}

\item{method}{Character indicating the method to be applied under the
common-item nonequivalent groups design (see Details). Options are
\code{"nominal weights"}, \code{"tucker"}, \code{"levine observed"}, \code{"levine true"}, and
\code{"chained"}.}

\item{circle}{If \code{NULL} (default), then circle-arc equating is not performed.
Else, options are \code{"simplified"} and \code{"symmetric"}.}

\item{...}{Additional arguments to be used with specific methods.
\describe{
\item{\code{weights}}{Use with nominal weights, Tucker, and Levine
observed-score equating. A numeric vector of length 2 indicating the
weights to be applied to the new and old populations, respectively. If
\code{NULL} (default), then weights are proportional to sample size.}
\item{\code{anchor}}{Use with Levine observed-score and Levine true-score
equating. An \code{"internal"} anchor (default) contributes to the total
test score, while an \code{"external"} anchor does not.}
\item{\code{lower}}{Use with circle-arc equating. A numeric vector of length 2
indicating the lower end points to be used for \code{x} and \code{y},
respectively. Default is \code{c(0, 0)}.}
}}

\item{verbose}{Logical indicating whether the full output should be returned.
Default is \code{TRUE}. If \code{FALSE}, then only the equated scores are returned.}
}
\value{
An object of class \code{eqt} is returned with the supplied arguments and
the following output.
\item{synth}{A matrix of the mean and SD of the synthetic populations, if
estimated.}
\item{coef}{A vector of coefficients.}
\item{scales}{A matrix of all possible raw scores and equated scores.}
\item{scores}{A matrix of raw scores and equated scores for each examinee.}
}
\description{
A collection of methods commonly used to perform small-sample
equating.
}
\details{
If the test forms have no items in common (\code{xv} or \code{yv = NULL}), then the
random groups design is assumed, and a method is not required.

If \code{xv} and \code{yv} are provided, then the common-item nonequivalent groups
design is assumed, and a method is applied to estimate the synthetic
population parameters. Depending on the method, additional arguments may be
required.
}
\examples{
# Random groups design ------------------------------------------------------

# Extract data
list2env(random_groups_dat, envir = globalenv())

# Identity equating
eqt(x, y)

# Mean equating
eqt(x, y, linear = "mean")

# Linear equating
eqt(x, y, linear = "linear")

# Common-item nonequivalent groups design -----------------------------------

# Extract data
list2env(nonequiv_groups_dat, envir = globalenv())

# Tucker mean equating
eqt(x, y, xv = 1:30, yv = 1:30, linear = "mean", method = "tucker")

# Levine observed-score equating with an external anchor
eqt(x, y, xv = 1:30, yv = 1:30, linear = "linear",
    method = "levine observed", anchor = "external")

# Simplified circle-arc with chained linear equating
eqt(x, y, xv = 1:30, yv = 1:30, linear = "linear", method = "chained",
    circle = "simplified")

}
\references{
Babcock, B., Albano, A., & Raymond, M. (2012). Nominal weights mean equating:
A method for very small samples. \emph{Educational and Psychological Measurement},
\emph{72}(4), 608--628. \url{https://doi.org/10.1177/0013164411428609}

Kolen, M. J., & Brennan, R. L. (2014). \emph{Test Equating, Scaling, and Linking:
Methods and Practices} (3rd ed.). Springer.
\url{https://doi.org/10.1007/978-1-4939-0317-7}

Livingston, S. A., & Kim, S. (2008). \emph{Small-sample equating by the circle-arc
method} (Research Report 08-39). Educational Testing Service.
}
\seealso{
\code{\link[=boot]{boot()}}, \code{\link[=multipleqt]{multipleqt()}}
}
\author{
Kylie N. Gorney \email{kyliengorney@gmail.com}
}
