% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-functions.R
\name{is.eqt}
\alias{is.eqt}
\alias{is.multipleqt}
\alias{is.boot}
\title{Is an object of a particular class?}
\usage{
is.eqt(...)

is.multipleqt(...)

is.boot(...)
}
\arguments{
\item{...}{One or more objects to be tested.}
}
\description{
Test whether an object is of a particular class.
}
\seealso{
\code{\link[=boot]{boot()}}, \code{\link[=eqt]{eqt()}}, \code{\link[=multipleqt]{multipleqt()}}
}
\author{
Kylie N. Gorney \email{kyliengorney@gmail.com}
}
