% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleqt.R
\name{multipleqt}
\alias{multipleqt}
\title{Weight multiple equating functions}
\usage{
multipleqt(
  ...,
  weights,
  method = c("point-wise", "angle bisector"),
  p = 2,
  verbose = TRUE
)
}
\arguments{
\item{...}{One or more objects of class \code{eqt}.}

\item{weights}{A vector of weights to be applied to the equatings. Default
weights all equatings equally.}

\item{method}{\code{"point-wise"} (default) uses the point-wise weighted average
method, and \code{"angle bisector"} uses the weighted angle bisector method (see
Details).}

\item{p}{Use when \code{method = "angle bisector"}. An integer greater than or
equal to 1 indicating the \eqn{L_p} circle to be used. Default is \code{2}.}

\item{verbose}{Logical indicating whether the full output should be returned.
Default is \code{TRUE}. If \code{FALSE}, then only the equated scores are returned.}
}
\value{
An object of class \code{multipleqt} is returned with the supplied
arguments and the following output. Equatings are numbered such that \code{eqt1}
is the first equating and so on.
\item{coef}{A matrix of coefficients.}
\item{scales}{A matrix of all possible raw scores and equated scores.}
\item{scores}{A matrix of raw scores and equated scores for each examinee.}
}
\description{
Weight multiple equating functions using the point-wise weighted
average method or the weighted angle bisector method.
}
\details{
If all equatings are parallel (i.e., they all have the same slope), then the
two methods will produce identical results. Else, the results will differ.
\itemize{
\item The point-wise weighted average method uses the exact \code{weights} that are
provided. However, the resulting transformation is not symmetric unless all
equatings are parallel. Circle-arc equatings are accepted as input.
\item The weighted angle bisector method adjusts the \code{weights} to ensure that the
resulting transformation is symmetric. However, circle-arc equatings are
not accepted as input.
}
}
\examples{
# Setup ---------------------------------------------------------------------

# Extract data
list2env(random_groups_dat, envir = globalenv())

# Identity equating
id <- eqt(x, y)

# Mean equating
me <- eqt(x, y, linear = "mean")

# Linear equating
li <- eqt(x, y, linear = "linear")

# Synthetic linking functions -----------------------------------------------

# Equal weights
multipleqt(id, li)

# Custom weights
multipleqt(id, li, weights = c(0.8, 0.2))

# Symmetric weights
multipleqt(id, li, method = "angle bisector")

# Other linking functions ---------------------------------------------------

# Use three equatings
multipleqt(id, me, li)

# Without identity equating
multipleqt(me, li)

}
\references{
Holland, P. W., & Strawderman, W. E. (2011). How to average equating
functions, if you must. In A. A. von Davier (Ed.), \emph{Statistical Models for
Test Equating, Scaling, and Linking} (pp. 89--107). Springer.
\url{https://doi.org/10.1007/978-0-387-98138-3_6}

Kolen, M. J., & Brennan, R. L. (2014). \emph{Test Equating, Scaling, and Linking:
Methods and Practices} (3rd ed.). Springer.
\url{https://doi.org/10.1007/978-1-4939-0317-7}
}
\seealso{
\code{\link[=boot]{boot()}}, \code{\link[=eqt]{eqt()}}
}
\author{
Kylie N. Gorney \email{kyliengorney@gmail.com}
}
