% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled.R
\name{scaled}
\alias{scaled}
\title{Compute scaled scores}
\usage{
scaled(x, point1 = mean(x), point2, sd2, digits = 0)
}
\arguments{
\item{x}{A numeric vector to be scaled.}

\item{point1, point2}{\code{point1} is a single score to be scaled to \code{point2}. For
example, if \code{point1 = mean(x)} (default), then \code{point2} represents the mean
of the new scale.}

\item{sd2}{The standard deviation of the new scale.}

\item{digits}{The number of decimal places to retain when rounding the scaled
scores. Default is \code{0}.}
}
\value{
A vector of scaled scores is returned.
}
\description{
Compute scaled scores by defining properties of the new scale.
}
\examples{
# Setup ---------------------------------------------------------------------

# Extract data
list2env(random_groups_dat, envir = globalenv())

# Raw test scores
raw <- rowSums(x)

# Compute scaled scores -----------------------------------------------------

# Define the new mean
scaled(raw, point2 = 500)

# Define the new mean and SD
scaled(raw, point2 = 500, sd2 = 100)

# Define the new maximum and SD
scaled(raw, point1 = max(raw), point2 = 1000, sd2 = 100)

}
\author{
Kylie N. Gorney \email{kyliengorney@gmail.com}
}
